Revert "static_cast to size_t"
This reverts commit c591c1ec29bcd831ac04765a0063b8bc05231c0f.

+++ Extensions/HSppConsole/Console.cpp
@@ -45,7 +45,7 @@ size_t getInputNum(const std::string& inputStr)
 
     if (isNumber(inputStr))
     {
-        size_t inputNum = static_cast<size_t>(std::stoi(inputStr));
+        int inputNum = std::stoi(inputStr);
         return inputNum;
     }
     else
